% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stute_test.R
\name{stute_test}
\alias{stute_test}
\title{stute_test}
\usage{
stute_test(
  df,
  Y,
  D,
  G = NULL,
  T = NULL,
  order = 1,
  seed = NULL,
  brep = 500,
  baseline = NULL
)
}
\arguments{
\item{df}{df}

\item{Y}{Y}

\item{D}{D}

\item{G}{G}

\item{T}{T}

\item{order}{order}

\item{seed}{seed}

\item{brep}{brep}

\item{baseline}{baseline}
}
\description{
stute_test
}
